+++ packages/promise-cache/src/promiseCache.ts
@@ -1,6 +1,6 @@
+import { type UUID, randomUUID } from 'node:crypto'
 import type { RedisClientOptions } from 'redis'
-import type { Persistor } from './persistor'
-import { createPersistor } from './persistor'
+import { Persistor } from './persistor'
 
 export type { RedisClientOptions }
 
@@ -8,14 +8,44 @@ export type PromiseCacheOptions = {
   ttlInSeconds?: number
   caseSensitive?: boolean
   redis?: RedisClientOptions
-  onError?: () => void
+  onError?: (error: string) => void
   onSuccess?: () => void
 }
 
-export const promises = {}
+const persistors: Record<string, Persistor> = {}
+
+const getPersistor = ({
+  redis,
+  onError,
+  onSuccess,
+  clientId,
+}: PromiseCacheOptions & { clientId: UUID }) => {
+  const connectionName = redis ? redis?.name || 'default' : 'local'
+
+  if (!persistors[connectionName]) {
+    persistors[connectionName] = new Persistor({
+      redis,
+      onError: (error: string) => {
+        onError?.(error)
+        console.error(
+          `❌ REDIS | Client Error | ${connectionName} | ${redis?.url}: ${error}`
+        )
+      },
+      onSuccess: () => {
+        onSuccess?.()
+        console.log(
+          `📦 REDIS | Connection Ready | ${connectionName} | ${redis?.url}`
+        )
+      },
+      clientId,
+    })
+  }
+  return persistors[connectionName]
+}
 
 export class PromiseCache<U> {
   public persistor: Persistor
+  private clientId: UUID = randomUUID()
   private readonly caseSensitive: boolean
   private readonly ttl?: number // Time to live in milliseconds.
 
@@ -31,7 +61,12 @@ export class PromiseCache<U> {
     onSuccess,
     onError,
   }: PromiseCacheOptions) {
-    this.persistor = createPersistor({ redis, onError, onSuccess })
+    this.persistor = getPersistor({
+      redis,
+      onError,
+      onSuccess,
+      clientId: this.clientId,
+    })
     this.caseSensitive = caseSensitive
     if (ttlInSeconds) {
       this.ttl = ttlInSeconds * 1000 // Convert seconds to milliseconds.
@@ -85,12 +120,14 @@ export class PromiseCache<U> {
    * @param key Cache key.
    * @param delegate The function to execute if the key is not in the cache.
    * @param ttlInSeconds Time to live in seconds.
+   * @param ttlKeyInSeconds The key in the response object that contains the TTL.
    * @returns The result of the delegate function.
    */
   async wrap(
     key: string,
     delegate: () => Promise<U>,
-    ttlInSeconds?: number
+    ttlInSeconds?: number,
+    ttlKeyInSeconds?: string
   ): Promise<U> {
     const now = Date.now()
 
@@ -98,13 +135,13 @@ export class PromiseCache<U> {
     const effectiveKey = this.caseSensitive ? key : key.toLowerCase()
 
     // Determine the TTL and unique cache key for this specific call.
-    const effectiveTTL =
+    let effectiveTTL =
       ttlInSeconds !== undefined ? ttlInSeconds * 1000 : this.ttl
 
     const cached = await this.persistor.get<U>(effectiveKey)
 
     if (cached) {
-      if (cached.ttl !== effectiveTTL) {
+      if (!ttlKeyInSeconds && cached.ttl !== effectiveTTL) {
         console.error(
           `WARNING: TTL mismatch for key: ${effectiveKey}. It is recommended to use the same TTL for the same key.`
         )
@@ -115,6 +152,14 @@ export class PromiseCache<U> {
 
     // Execute the delegate, cache the response with the current timestamp, and return it.
     const response = await delegate()
+
+    // Get the TTL from the response if a TTL key is provided.
+    if (ttlKeyInSeconds) {
+      const responseDict = response as Record<string, unknown>
+      const responseTTL = Number(responseDict[ttlKeyInSeconds] as string) * 1000
+      effectiveTTL = responseTTL || effectiveTTL // Fall back to the default TTL if the TTL key is not found.
+    }
+
     this.persistor.set(effectiveKey, {
       value: response,
       timestamp: now,